Move config components into a subfolder
